backend/requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
colorama==0.4.6
Django==5.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
flake8==7.1.1
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.7.0
mypy==1.13.0
mypy-extensions==1.0.0
mysqlclient==2.2.6
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
setuptools==75.6.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
wheel==0.45.1
